pyproject.toml
[tool.poetry]
name = "pyderground"
version = "0.1.0"
description = "VX Underground file processor with YARA scan"
authors = ["PalPat1ne Gen1us <serg@12345290@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.18.0"
yara-python = "^4.1.3"
requests = "^2.26.0"
py7zr = "^0.22.0"


